requirements.txt
asgiref==3.8.1
Django
django-cors-headers
django-sslserver
djangorestframework
gunicorn==23.0.0
packaging==24.1
PyMySQL==1.1.1
sqlparse==0.5.1
django-filter
cryptography
django-crontab
python-dotenv
pillow
pandas
openpyxl